feat: implement BoxPlotPanelRenderer and supporting utilities for box plot chart rendering
